backend/requirements.txt
pandas
requests
python-dotenv
supabase
pdfplumber
numpy
scikit-learn